Moved the get_provider() function from ocp to node module
